changed datatype to accomodate for 0.x multipliers

+++ include/dmg_calculation.h
@@ -3,12 +3,14 @@
 
 #include "inputhandler.h"
 
-namespace dmgCalc {
+namespace dmgCalc
+{
     int crit_dmg(int, float);
     int em_perc_bonus(int);
     int basic_atk(int, float, int);
-    int outgoing_dmg(int, int, int);
+    int outgoing_dmg(int, float, float);
     int amping_dmg(int, float, int);
+
 };
 
 #endif
